Created Code for Generating Cards.
Created a script for cardData for creation of Scriptable objects.

Added Different Scriptable Objects for different Cards, such as:-
* FireAttackCard
*Fire2AttackCard
*IceAttackCard
*Ice2AttackCard
*FireandIceCard
*Health1Card
*Health2Card
*MirrorCard;

Then created a Card Script to keep different data for cards and attached to card object in Hierarchy.

Created arrays named DamageNumbers and HealthNumbers as they were sprites and needed to be referenced as numbers when required.

Also assigned each sprites with their respective numbers in arrays.